requirements.txt
grpcio
grpcio-tools
typer
